chrono fini

+++ Assets/Scripts/Scoring/PlayerScoring.cs
@@ -3,5 +3,5 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class PlayerScoring {
-	public int highScore;
+	public float highScore;
 }
